Implemented UVAArenaWidget and modified subclasses
All subclasses of UVAArenaWidget can access each other via the appropriate
accessor method. Each subclass can also emit a signal that lets other
subclasses be aware of a change. Expensive operations should not be done
in the UVAArenaEvent signals.

Updated the MainWindow class and the ProblemsWidget class, and also
updated the Uhunt api problem list download so that it returns a QMap
instead of a QList. Removed the UhuntDatabase class.